Added some styling to the Djs Dashboard.
