+++ src/css/index.css
@@ -1,0 +1,24 @@
+.app {
+  background-color: #383838;
+  color: white;
+}
+.dashboard {
+  display: flex;
+  flex-flow: column nowrap;
+  align-items: center;
+}
+.dashboard h1 {
+  margin-top: 5vh;
+}
+.dashboard .board {
+  display: flex;
+  flex-flow: row nowrap;
+  justify-content: space-between;
+  width: 40%;
+  border: solid #333A42 1px;
+  border-radius: 25px;
+  padding: 5%;
+}
+.dashboard .welcome {
+  margin-top: 3vh;
+}
