+++ src/less/app.less
@@ -1,0 +1,4 @@
+.app {
+    background-color: #383838;
+    color: white;
+}
