+++ src/less/dashboard.less
@@ -1,0 +1,23 @@
+.dashboard {
+    h1 {
+        margin-top: 5vh;
+    }
+    display: flex;
+    flex-flow: column nowrap;
+    align-items: center;
+
+    .board {
+        display: flex;
+        flex-flow: row nowrap;
+        justify-content: space-between;
+        width: 40%;
+        border: solid @dashboardColor 1px;
+        border-radius: 25px;
+        padding: 5%;
+    }
+
+    .welcome {
+        margin-top: 3vh;
+    }
+
+}
