+++ src/less/variables.less
@@ -1,0 +1,1 @@
+@dashboardColor: #333A42; //The Reactstrap dark mode color
